Rename `utils` dir to `lib`, remove index
