+++ src/components/general/IconButton/index.tsx
@@ -5,7 +5,7 @@ import React, {
 } from 'react';
 import { Primitive } from '@radix-ui/react-primitive';
 
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 
 export const IconButton = forwardRef<
   HTMLButtonElement,
