+++ src/components/general/List/ListItem.tsx
@@ -10,7 +10,7 @@ import {
   useRovingFocusGroupScope,
   type ScopedProps,
 } from './common';
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 
 // eslint-disable-next-line react/display-name
 const ListItemButton = forwardRef<
