+++ src/components/general/List/ListItemText.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, type ComponentPropsWithoutRef } from 'react';
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 
 export interface ListItemTextProps extends ComponentPropsWithoutRef<'div'> {
   primary?: string;
