+++ src/components/general/List/ListSubheader.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, type ComponentPropsWithoutRef } from 'react';
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 
 export const ListSubheader = forwardRef<
   HTMLSpanElement,
