+++ src/components/general/ScrollArea/ScrollArea.tsx
@@ -11,7 +11,7 @@ import {
   ScrollAreaCorner,
 } from '@radix-ui/react-scroll-area';
 
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 import { ScrollBar } from './ScrollBar';
 
 export const ScrollArea = forwardRef<
