+++ src/components/general/Sheet/Sheet.tsx
@@ -8,7 +8,7 @@ import {
   DialogPortal,
   type DialogProps,
 } from '@radix-ui/react-dialog';
-import { cn } from 'utils/cn';
+import { cn } from 'lib/cn';
 
 export const Sheet = forwardRef<
   HTMLDivElement,
