+++ src/components/layout/Mdx/Mdx.tsx
@@ -7,7 +7,7 @@ import { Alert } from './Alert';
 import { MdxBlockquote } from './MdxBlockquote';
 import { MdxLink } from './MdxLink';
 import * as layouts from 'layouts';
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 import type { Doc } from 'types';
 
 export const mdxComponents: MDXComponents = {
