+++ src/components/layout/Search/SearchBoxButton.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, type ComponentPropsWithoutRef } from 'react';
 
 import { SearchIcon } from 'assets/icons';
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 
 export const SearchBoxButton = forwardRef<
   HTMLButtonElement,
