+++ src/components/layout/Sidebar/SidebarItem.tsx
@@ -2,7 +2,7 @@
 
 import { useMemo } from 'react';
 import Link from 'next/link';
-import { cn } from 'utils';
+import { cn } from 'lib/cn';
 
 import { ListItem, ListItemIcon, ListItemText } from 'components/general';
 import { ChevronRightIcon } from 'assets/icons';
